[Bronze III] Title: 수학적 호기심, Time: 76 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/9094. 수학적 호기심/수학적 호기심.cc
@@ -1,5 +1,4 @@
 #include <iostream>
-#include <string>
 using namespace std;
 
 int main() {
@@ -8,12 +7,12 @@ int main() {
     int T; cin >> T;
 
     while(T--) {
-        double n, m; cin >> n >> m;
+        int n, m; cin >> n >> m;
         int cnt = 0;
 
-        for(double a=1; a<n-1; a++) {
-            for(double b=a+1; b<n; b++) {
-                double e = (a*a+b*b+m)/(a*b);
+        for(int a=1; a<n-1; a++) {
+            for(int b=a+1; b<n; b++) {
+                double e = (double)(a*a+b*b+m)/(a*b);
                 if(e == (int)e) cnt++;
             }
         }
